Surfaces context sensitive menu and counts (plus bug fixes)

+++ registry2/trunk/registry-ws/src/main/java/org/gbif/registry2/ws/resources/DatasetResource.java
@@ -385,7 +385,7 @@ public class DatasetResource extends BaseNetworkEntityResource<Dataset>
   @GET
   @Override
   public PagingResponse<Dataset> listConstituents(@PathParam("key") UUID datasetKey, @Context Pageable page) {
-    return pagingResponse(page, datasetMapper.countConstituents(datasetKey),
+    return pagingResponse(page, (long) datasetMapper.countConstituents(datasetKey),
       datasetMapper.listConstituents(datasetKey, page));
   }
